+++ Command/ListEncodingsCommand.php
@@ -60,7 +60,7 @@ class ListEncodingsCommand extends CloudCommand
     /**
      * {@inheritDoc}
      */
-    public function execute(InputInterface $input, OutputInterface $output)
+    protected function doExecuteCommand(InputInterface $input, OutputInterface $output)
     {
         $filter = array();