+++ Command/ListProfilesCommand.php
@@ -35,7 +35,7 @@ class ListProfilesCommand extends CloudCommand
     /**
      * {@inheritDoc}
      */
-    public function execute(InputInterface $input, OutputInterface $output)
+    protected function doExecuteCommand(InputInterface $input, OutputInterface $output)
     {
         $profiles = $this->getCloud($input)->getProfiles();