+++ Command/UploadVideoCommand.php
@@ -41,7 +41,7 @@ class UploadVideoCommand extends CloudCommand
     /**
      * {@inheritDoc}
      */
-    public function execute(InputInterface $input, OutputInterface $output)
+    protected function doExecuteCommand(InputInterface $input, OutputInterface $output)
     {
         $output->writeln('Starting file upload...');
         $this->getCloud($input)->encodeVideoFile($input->getArgument('filename'));